+++ src/Enums/StepType.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Enadstack\Approvio\Enums;
+
+enum StepType: string
+{
+    case Sequential = 'sequential';
+    case Parallel = 'parallel';
+}